move, clean up, and document API resources
